Uploading the files like data science file
Two files of data science for example one source code file and the other is document source code file. This task is about to supervised machine learning and in this task, predict the percentage of student marks based on the number of study hours. For this purpose, use python and Google Colaboratory tool.